Fix interface for nextInt to return int instead of double

+++ ex0/src/to/richard/IRandom.java
@@ -18,7 +18,7 @@ public interface IRandom {
      * @param n
      * @return
      */
-    public double nextInt(int n);
+    public int nextInt(int n);
 
     /**
      * Returns the next pseudorandom, uniformly distributed double value between 0.0 and 1.0 from this random number generator's sequence.
